tools/requirements.txt
google-cloud-storage
jsonpickle
